app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==23.0.0
python-multipart==0.0.17
jinja2==3.1.4
Pillow==11.3.0
pdf2image==1.17.0
rapidfuzz==3.10.1
python-dotenv==1.0.1
httpx==0.28.1 
fuzzywuzzy==0.18.0
pypdf==5.1.0
# pytesseract==0.3.13 # Not used - removed
# python-Levenshtein==0.27.1 # Not used - fuzzywuzzy uses rapidfuzz backend

# Authentication
google-auth==2.37.0
google-auth-oauthlib==1.2.1
python-jose[cryptography]==3.3.0

# For testing
pytest==8.3.4
pytest-asyncio==0.25.0
# fpdf2==2.8.1 # Not used - removed